requirements.txt

pytest~=7.2.0
Django==4.1.3
djangorestframework==3.14.0
drf-yasg==1.21.4